fix) 인기 등록상품 api 추가

+++ gofield-domain/gofield-domain-rds/src/main/java/com/gofield/domain/rds/entity/itemBundleAggregation/ItemBundleAggregation.java
@@ -27,7 +27,7 @@ public class ItemBundleAggregation extends BaseTimeEntity {
     private int reviewCount;
 
     @Column
-    private int reviewScore;
+    private Double reviewScore;
 
     @Column
     private int newLowestPrice;
